feat: lint output per line
Adds an '-line' or '-l' option to the 'lint' command. Prints a line for
each error, starting with the file to which the error belongs.

This output can be a starting point for some parsers, in particular ones
that implement something similar to Vim errorformat parsing.

Related to https://github.com/fsfe/reuse-tool/issues/925

Additional work needed:

- [ ] Needs tests
- [ ] Error messages might have to be improved
- [ ] Not all errors are found, as some issues aren't in the
  FileReports. This requires additional investigation.

Signed-off-by: Nico Rikken <nico@nicorikken.eu>